requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1 
python-docx==1.1.0
openpyxl==3.1.2
# Cosmos DB接続用
pymongo>=4.0.0
# MySQL接続用
PyMySQL==1.1.0
# ベクトル検索用
rank-bm25
nltk
numpy<2.0.0
scikit-learn
# Gremlin接続用
gremlinpython==3.7.1
